LayoutTests/platform/wk2/http/tests/contentextensions/block-everything-unless-domain-expected.txt

CONSOLE MESSAGE: helper loaded
CONSOLE MESSAGE: Content blocker prevented frame displaying http://localhost:8000/contentextensions/resources/should-not-load.html from loading a resource from http://localhost:8000/contentextensions/resources/should-not-load.html
CONSOLE MESSAGE: PASS - blocked the load
